Enforce and use std::int64_t as time datatype

+++ gameserver/utils/export/tick.h
@@ -30,7 +30,7 @@
 class Tick
 {
  public:
-  static std::uint32_t now();
+  static std::int64_t now();
 };
 
 #endif  // UTILS_EXPORT_TICK_H_
